refactor(audio): deduplicate cross-platform audio device, system audio, and cpal streams
