+++ src-tauri/src/audio/pipeline/input/windows.rs
@@ -1,18 +1,13 @@
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
-use serde_json::json;
-use tauri::{AppHandle, Emitter};
-use tracing::{error, info};
+use tauri::AppHandle;
+use tracing::info;
 
-use crate::audio::device::{self, DeviceKind};
 use crate::audio::graph::{InputSpec, ValidInput};
-use crate::audio::health;
 use crate::audio::input_bridge::BroadcastRx;
-use crate::audio::streams;
 use crate::error::AppResult;
 
-use super::super::native::native_config;
 use super::{resolve_audio_file, start_audio_file, InputHandle, ResolvedInput};
 
 const LOOPBACK_FALLBACK_RATE: u32 = 48_000;
@@ -21,17 +16,7 @@ const LOOPBACK_CHANNELS: usize = 2;
 
 pub(in crate::audio::pipeline) fn resolve_input(inp: &ValidInput) -> AppResult<ResolvedInput> {
     match &inp.spec {
-        InputSpec::Microphone { device_id } => {
-            let device = device::find(DeviceKind::Input, device_id)?;
-            let native = native_config(DeviceKind::Input, &device, device_id)?;
-            Ok(ResolvedInput::Cpal {
-                device,
-                config: native.config,
-                sample_format: native.sample_format,
-                src_channels: native.channels as usize,
-                sample_rate: native.sample_rate,
-            })
-        }
+        InputSpec::Microphone { device_id } => super::cpal::resolve_cpal_input(device_id),
         InputSpec::SystemAudio {
             exclude_current_app,
         } => Ok(ResolvedInput::SystemAudio {
@@ -65,33 +50,16 @@ pub(in crate::audio::pipeline) fn start_input_stream(
             sample_format,
             src_channels,
             ..
-        } => {
-            let dead = Arc::new(AtomicBool::new(false));
-            let dead_cb = dead.clone();
-            let app_err = app.clone();
-            let node_id_cb = node_id.to_string();
-            let err_cb = move |e: cpal::StreamError| {
-                if dead_cb.swap(true, Ordering::Relaxed) {
-                    return;
-                }
-                health::bump(&health::STREAM_ERRORS, 1);
-                error!(node_id = %node_id_cb, error = %e, "input stream error");
-                let _ = app_err.emit(
-                    "audio://input_error",
-                    json!({ "nodeId": node_id_cb, "error": format!("{e}") }),
-                );
-            };
-            let stream = streams::build_input_stream(
-                &device,
-                &config,
-                sample_format,
-                src_channels,
-                bridge,
-                meter,
-                err_cb,
-            )?;
-            Ok(InputHandle::Cpal(stream))
-        }
+        } => super::cpal_input::start_cpal_input_stream(
+            node_id,
+            device,
+            config,
+            sample_format,
+            src_channels,
+            bridge,
+            meter,
+            app,
+        ),
         ResolvedInput::SystemAudio {
             sample_rate,
             exclude_current_app,
