+++ src-tauri/src/audio/pipeline/output/macos.rs
@@ -3,67 +3,24 @@ use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
 
-use cpal::traits::{DeviceTrait, StreamTrait};
+use cpal::traits::DeviceTrait;
 use serde_json::json;
 use tauri::{AppHandle, Emitter};
 use tracing::{error, info, warn};
 
-use crate::audio::device::{self, DeviceKind};
 use crate::audio::health;
 use crate::audio::streams;
 use crate::error::{AppError, AppResult};
 
 use super::super::dag::OutputGraph;
-use super::super::native::native_config;
 use super::super::worker::WorkerCtrl;
-use super::{spawn_speaker_worker, speaker_ring, SpeakerIo, SpeakerWorker, StreamGuard};
+use super::{spawn_speaker_worker, speaker_ring, SpeakerIo, StreamGuard};
 
 // Bluetooth AUHAL often returns DeviceNotAvailable on first bind; retry covers settling.
 const SPEAKER_MAX_ATTEMPTS: u32 = 3;
 const SPEAKER_RETRY_DELAY: Duration = Duration::from_millis(300);
 
-pub(in crate::audio::pipeline) struct SpeakerResolved {
-    pub device: cpal::Device,
-    pub config: cpal::StreamConfig,
-    pub sample_format: cpal::SampleFormat,
-    pub out_channels: usize,
-    pub sample_rate: u32,
-}
-
-// Field order: the stream drops before the worker so the audio callback stops
-// before the ring is freed.
-pub(in crate::audio::pipeline) struct SpeakerHandle {
-    _stream: cpal::Stream,
-    _worker: SpeakerWorker,
-    _alive: StreamGuard,
-}
-
-// cpal's coreaudio backend registers a device-alive property listener for any
-// non-default device (which `device::find` always returns -- see its comment)
-// whose callback closure holds another clone of the `Stream`'s inner `Arc`.
-// That's a permanent reference cycle: dropping our `_stream` handle alone
-// never reaches refcount zero, so the AudioUnit is never disposed and keeps
-// calling `fill` on a ring nobody drains anymore. `pause` reaches the
-// AudioUnit through `&self` and stops it for real, independent of the cycle.
-impl Drop for SpeakerHandle {
-    fn drop(&mut self) {
-        if let Err(e) = self._stream.pause() {
-            warn!(error = %e, "failed to pause speaker stream on teardown");
-        }
-    }
-}
-
-pub(in crate::audio::pipeline) fn resolve_speaker(device_id: &str) -> AppResult<SpeakerResolved> {
-    let device = device::find(DeviceKind::Output, device_id)?;
-    let native = native_config(DeviceKind::Output, &device, device_id)?;
-    Ok(SpeakerResolved {
-        device,
-        config: native.config,
-        sample_format: native.sample_format,
-        out_channels: native.channels as usize,
-        sample_rate: native.sample_rate,
-    })
-}
+pub(in crate::audio::pipeline) use super::cpal_speaker::{resolve_speaker, SpeakerHandle, SpeakerResolved};
 
 // Substring match on cpal's stable Display -- AppError flattens the variant.
 fn is_device_not_available(e: &AppError) -> bool {
@@ -178,11 +135,7 @@ pub(in crate::audio::pipeline) fn start_speaker_stream(
         meter,
     )?;
     Ok((
-        SpeakerHandle {
-            _stream: stream,
-            _worker: worker_handle,
-            _alive: StreamGuard::new(),
-        },
+        SpeakerHandle::new(stream, worker_handle, StreamGuard::new()),
         ctrl,
         dead,
         io,
